pipemgr/pyproject.toml
[tool.poetry]
name = "pipemgr"
version = "1.0.0"
description = "Helps manage the book syncing pipeline"
authors = ["Tyler Nullmeier <tylerzeromaster@gmail.com>", "m1yag1 <mike.arbelaez@rice.edu>"]
packages = [
    { include = "pipemgr", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.8"
PyYAML = "^6.0"
docopts = "^0.6.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pydantic = "^1.8.2"

[tool.poetry.scripts]
pipemgr = "pipemgr.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
